Patch: restructure and dockerfiles
* A bit o' restructuring - it was required.
* Stole the license check from Unikorn - Thanks @spjmurray 😉 
* Updated the place to store the results file for scan and build.
* Added Dockerfile - this will require a manual building of the binary
until the Makefile is added (coming soon to a pull request near you)

The results adjustment is required for the Argo workflows work as it's
failing to find the fetched `results.json` file thanks to me putting
relative pathing in place originally. \o/

Once this is merged I'll bump the action and auto-kube-image tags as
required - probably will skip the PR as it's not going to change
anywhere other than the tagged version.

Fundamentally, nothing is changing really, just some additions, all of
which have been checked and the restructure - again which has been
checked.