feat: Added new image signing command
Image signing can now be done via Baski. It'll take the image name, hadh it and then sign it using a provided private key (or one from vault if credentials supplied)
The sign command supports generating a new public/private keypair, signing and image and also verifying the digest so that image signing can be confirmed.