fix: changed fip selection to fip creation to prevent race condition
This create a new process for getting Floating IPs onto the server in
OpenStack. The previous approach looked for FIPs in the account and if
it couldn't find one, would fail. The problem with this was if three
scans ran at the same time there would be a fight over who got the IP
and multiple servers may try to attach it at the same time. OpenStack
wouldn't allow this but Baski was ignorant to this and so would try to
continue.

Now a floating IP is allocated to the project as they are required
preventing any race condition and it is released back into the pool
after.

This PR also contains some improvements around how to detect errors
during the process and separates some logic out such as creating a
server and attaching an IP to it into separate processes to allow better
detection of errors and to keep it SOLID.

Fixes #38