feat: trivyignore support
This PR contains code to allow a user to pass in options for either
downloading a file, adding a list of CVEs or both where the list is
concatenated onto the existing file temporarily.

Also added extensive guidance to the baski-example file to help users
understand each option available to them.

Updated some Go Packages.